fix: satisfy code index lint gates
