+++ src/cli_v1_routes_d.c
@@ -1450,8 +1450,7 @@ static int cli_ws_ingest_root(const char *remote, const char *bearer, const char
       return 1;
    }
 
-   if (cli_ws_ingest_phase(remote, bearer, base, abs_root, "seal", s.scan_id, s.pushed, NULL) !=
-       0)
+   if (cli_ws_ingest_phase(remote, bearer, base, abs_root, "seal", s.scan_id, s.pushed, NULL) != 0)
    {
       (void)cli_ws_ingest_phase(remote, bearer, base, abs_root, "abort", s.scan_id, 0, NULL);
       return 1;
